refactor(booking_card): rewrite Widgets to allow a more flexible composition
